+++ src/libtrx/game/game_flow/reader.c
@@ -36,6 +36,8 @@ typedef struct {
     void *handler_func_arg;
 } M_SEQUENCE_EVENT_HANDLER;
 
+static bool M_ParseRGB888(JSON_VALUE *value, RGB_888 *target);
+
 static M_SEQUENCE_EVENT_HANDLER *M_GetSequenceEventHandlers(void);
 
 typedef void (*M_LOAD_ARRAY_FUNC)(
@@ -95,6 +97,35 @@ static void M_LoadRoot(const M_CONTEXT *ctx, JSON_OBJECT *obj);
     #include "./reader_tr2.def.c"
 #endif
 
+static bool M_ParseRGB888(JSON_VALUE *const value, RGB_888 *const target)
+{
+    if (value != nullptr && value->type == JSON_TYPE_ARRAY) {
+        const JSON_ARRAY *const tmp_arr = JSON_ValueAsArray(value);
+        const RGB_F color = {
+            JSON_ArrayGetDouble(tmp_arr, 0, -1.0),
+            JSON_ArrayGetDouble(tmp_arr, 1, -1.0),
+            JSON_ArrayGetDouble(tmp_arr, 2, -1.0),
+        };
+        if (color.r >= 0.0 && color.g >= 0.0 && color.b >= 0.0) {
+            *target = (RGB_888) {
+                color.r * 255.0f,
+                color.g * 255.0f,
+                color.b * 255.0f,
+            };
+            return true;
+        }
+    } else if (value != nullptr && value->type == JSON_TYPE_STRING) {
+        const char *tmp_str = JSON_ValueGetString(value, JSON_INVALID_STRING);
+        ASSERT(tmp_str != JSON_INVALID_STRING);
+        RGB_888 tmp_color;
+        if (String_ParseRGB888(tmp_str, &tmp_color)) {
+            *target = tmp_color;
+            return true;
+        }
+    }
+    return false;
+}
+
 static void M_LoadCommonSettings(
     const M_CONTEXT *const ctx, JSON_OBJECT *const obj,
     GF_LEVEL_SETTINGS *const settings)
@@ -118,32 +149,25 @@ static void M_LoadCommonSettings(
     }
 
     {
+        const int value =
+            JSON_ObjectGetBool(obj, "fog_transparency", JSON_INVALID_BOOL);
+        if (value != JSON_INVALID_BOOL) {
+            settings->fog_transparency.is_present = true;
+            settings->fog_transparency.value = value;
+        }
+    }
+
+    {
+        JSON_VALUE *const tmp_value = JSON_ObjectGetValue(obj, "fog_color");
+        if (M_ParseRGB888(tmp_value, &settings->fog_color.value)) {
+            settings->fog_color.is_present = true;
+        }
+    }
+
+    {
         JSON_VALUE *const tmp_value = JSON_ObjectGetValue(obj, "water_color");
-        if (tmp_value != nullptr && tmp_value->type == JSON_TYPE_ARRAY) {
-            const JSON_ARRAY *const tmp_arr = JSON_ValueAsArray(tmp_value);
-            const RGB_F color = {
-                JSON_ArrayGetDouble(tmp_arr, 0, -1.0),
-                JSON_ArrayGetDouble(tmp_arr, 1, -1.0),
-                JSON_ArrayGetDouble(tmp_arr, 2, -1.0),
-            };
-            if (color.r >= 0.0 && color.g >= 0.0 && color.b >= 0.0) {
-                settings->water_color.is_present = true;
-                settings->water_color.value = (RGB_888) {
-                    color.r * 255.0f,
-                    color.g * 255.0f,
-                    color.b * 255.0f,
-                };
-            }
-        } else if (
-            tmp_value != nullptr && tmp_value->type == JSON_TYPE_STRING) {
-            const char *tmp_str =
-                JSON_ValueGetString(tmp_value, JSON_INVALID_STRING);
-            ASSERT(tmp_str != JSON_INVALID_STRING);
-            RGB_888 tmp_color;
-            if (String_ParseRGB888(tmp_str, &tmp_color)) {
-                settings->water_color.is_present = true;
-                settings->water_color.value = tmp_color;
-            }
+        if (M_ParseRGB888(tmp_value, &settings->water_color.value)) {
+            settings->water_color.is_present = true;
         }
     }