+++ src/libtrx/include/libtrx/config/types_tr2.h
@@ -58,6 +58,8 @@ typedef struct {
         float brightness;
 
         RGB_888 water_color;
+        bool fog_transparency;
+        RGB_888 fog_color;
         int32_t fog_start;
         int32_t fog_end;
     } visuals;